requirements.txt
# pip install pytorch==2.2.2 torchvision==0.17.2 torchaudio==2.2.2 pytorch-cuda=12.1 -c pytorch -c nvidia
pydantic
opencv-python
tqdm
matplotlib
albumentations
numpy==1.26.4
segmentation-models-pytorch>=0.3.0
timm>=0.9.2
